Separate walking a frame trace from printing the trace
Lock analysis requires getting a frame trace at the point the lock is taken,
but walking the frame trace is only used in printing out the trace. This
change adds a new machine function get_frame_ra_ptr to retrieve a pointer
to the return address for a given frame as well as return a pointer to the
next frame up. Since the printing of the frame trace is identical across
platforms except for walking the frame trace, the printing code is now
platform independent and moved to kernel.c.