Add support for lock profiling
This change introduces a new option to compile in lock profiling support.
Lock profiling added with a make variable, and is controlled runtime via an
http interface. Lock statistics are tracked per cpu and then collated together
when fetched via the http interface. A new python script can be used to analyze
the profile data to aid analysis.

To avoid infinite recursive locking, a new field on the cpuinfo controls
whether or not profiling is enabled allowing the lock profile code to acquire
locks without profiling them. Preallocated versions of an objcache and table
are used on each cpu to track statistics without the need for additional
locking and to improve performance when profiling is enabled.