aarch64: fix resuming of kernel and syscall contexts
When a kernel or syscall context is suspended via
context_suspend(), no hardware exception is raised, therefore
execution of an `eret` instruction when resuming the context in
frame_return() wrongly changes the CPU exception level from EL1 to
EL0, i.e. moves from kernel space to user space.
This change fixes the above issue by chaging the implementation of
context_suspend() so that the processor state is saved via the nzcv
register (instead of spsr_el1) and a zero value is put in the ELR
location of the saved frame (instead of using the elr_el1
register). frame_return() checks the ELR value in the frame being
restored: if zero, the processor state is restored via the nzcv
register and execution jumps to the next instruction of the
suspended context via a simple `ret` instruction (instead of
`ret`).